requirements.txt
astropy==5.3.4
dash==2.15.0
gymnasium==0.29.1
imageio==2.34.0
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.4
pandas==2.2.0
pettingzoo==1.24.3
Pillow==10.2.0
plotly==5.18.0
poliastro==0.17.0
PyYAML==6.0.1
scipy==1.12.0
stable_baselines3==2.2.1
SuperSuit==3.9.2
tensorflow>=2.15.0
torch==2.2.0
tqdm==4.66.2
pygame==2.5.2
pymunk==6.6.0